Adding test to verify the RegexStringCleaner using fuzzing files.
Parameterized test to read in the relative files and pass each unique
file entry to the test for validation.  Right now our expectation is
that an unchanged string represents a failed cleaning, but we will need
to work with our data set to make the cases valid.

For now, the test does fail.  Need to confer with Matt about trimming
the data set to correct expectations.